# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin API 文档",
        "title": "Gin Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "25076778@qq.com"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8080",
    "paths": {
        "/api/department": {
            "get": {
                "description": "部门列表",
                "tags": [
                    "部门管理"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "是否不分页",
                        "name": "notPage",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/request.PageData"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.Department"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "部门创建",
                "tags": [
                    "部门管理"
                ],
                "summary": "创建",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DepartmentCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Department"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/department/{id}": {
            "get": {
                "description": "部门详情",
                "tags": [
                    "部门管理"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Department"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "部门更新",
                "tags": [
                    "部门管理"
                ],
                "summary": "更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DepartmentUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/errcode.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "部门删除",
                "tags": [
                    "部门管理"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/errcode.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errcode.ArgsErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Invalid arguments"
                }
            }
        },
        "errcode.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "errcode.SystemErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
...
        "model.Department": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {}
                },
                "createdAt": {
                    "type": "string"
                },
                "deptLeaders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DepartmentLeaders"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.DepartmentLeaders": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "departmentId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "leader": {
                    "$ref": "#/definitions/model.User"
                },
                "leaderUserId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "fullName": {
                    "description": "姓名",
                    "type": "string"
                },
                "gender": {
                    "description": "性别 1=男 2=女",
                    "type": "integer"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "mainDept": {
                    "$ref": "#/definitions/model.UserDepartments"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "status": {
                    "description": "状态 1=启用 2=停用",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "userDepts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserDepartments"
                    }
                },
                "userRoles": {
                    "description": "用户角色",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserRoles"
                    }
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "model.UserDepartments": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "departmentId": {
                    "type": "integer"
                },
                "dept": {
                    "$ref": "#/definitions/model.Department"
                },
                "id": {
                    "type": "integer"
                },
                "isMain": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "model.UserRoles": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "roleId": {
                    "description": "角色id",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "userId": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
...
        "request.DepartmentCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "request.DepartmentUpdate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "request.PageData": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "数据列表"
                },
                "notPage": {
                    "description": "不分页",
                    "type": "boolean",
                    "example": false
                },
                "page": {
                    "description": "当前页",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
...
    }
}